generated_bots/bot_4.py
 Sure, here is an example of a Python code for a Discord bot using the microsoft/phi-4 model to chat and share knowledge about coding. This code assumes that you have already set up a Discord bot and have its API keys, which can be obtained by creating a new bot on the Discord Developer Portal.
```python
import discord
import phi_4

client = discord.Client()
model = phi_4.PhiModel()

@client.event
async def on_ready():
    print(f'We have logged in as {client.user}')
    
@client.event
async def on_message(message):
    if message.author == client.user:
        return
    
    # Check if the user has sent a coding-related question
    if "coding" in message.content.lower():
        # Generate a response using the Phi model
        response = model.generate_response(message.content.lower())
        
        # Send the generated response to the user's channel
        await message.channel.send(response)
    
client.run('YOUR_DISCORD_API_KEY')
```
This code uses the `discord` library to interact with Discord, and the `phi_4` library to use the Phi model for generating responses to coding-related questions. The `on_ready()` event handler is called when the bot has successfully logged in to Discord, while the `on_message()` event handler is called whenever a new message is sent to the bot's server. If the user's message contains the word "coding", the bot will use the Phi model to generate a response and send it back to the channel where the original message was sent. 