requirements.txt
#
# Requirements for Owlmind

discord==2.3.2
python-dotenv>=1.0.0
requests>=2.28.0
audioop-lts; python_version>='3.13'
faiss-cpu>=1.8.0  # Vector similarity search
Flask>=3.0.0  # Web server
PyPDF2>=3.0.0  # PDF processing
python-docx>=0.8.11  # Word document processing
docx2txt>=0.8  # Alternative text extraction
ollama>=0.1.6  # Ollama client integration
sentence-transformers>=2.6.0  # Text embeddings
numpy>=1.26.0  # Numerical processing for FAISS

